+++ src/main/java/com/projeto_inicial/projeto_inicial/Farm/Repository/PlotRepository.java
@@ -1,6 +1,6 @@
-package com.projeto_inicial.projeto_inicial.Plot.Repository;
+package com.projeto_inicial.projeto_inicial.Farm.Repository;
 
-import com.projeto_inicial.projeto_inicial.Plot.Model.Plot;
+import com.projeto_inicial.projeto_inicial.Farm.Model.Plot;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
